==> Udemy/PandasBootcamp/Course_Materials_Part2/Video_Lecture_NBs/NB_09_from_scratch.ipynb <==
import pandas as pd

sales = pd.read_excel("sales.xls", index_col=0)

sales

sales.info()

pd.read_excel('sales.xls', index_col=0, header = 0)

pd.read_excel('sales.xls', index_col = 0, names = ['name', 'city', 'nation', 'revn', 'adtl_comp'])
