==> Udemy/PandasBootcamp/Course_Materials_Part2/Video_Lecture_NBs/NB_09_Importing Excel Files.ipynb <==
# Part 2: Full Data Workflow A-Z

## Importing from Excel Files with pd.read_excel()

Before you can move on with the next Lectures on Importing Data from Excel with pd.read_excel() you may need to import xlrd with the following command (to be entered into __Anaconda Prompt__ / __Terminal Window__):

conda install xlrd

### First Steps

import pandas as pd

sales = pd.read_excel("sales.xls", index_col= 0)

sales

sales.info()

pd.read_excel("sales.xls", index_col = 0, header = 0)

pd.read_excel("sales.xls", index_col = 0, header = 0, names = ["Name", "Loc_City", "Loc_Country", "Revenue", "Add_Comp"])

pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = "A:C")

pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = "C:E")

pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = "A, C:E")

pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = ":C")

#pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = "C:")

pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = [0,3,4])

#pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = 2)

pd.read_excel("sales.xls", index_col = 0, header = 0, usecols = ["City", "Sales"])

### Customizing import with pd.read_excel()

import pandas as pd

pd.read_excel("summer_raw.xls")

pd.read_excel("summer_raw.xls", sheet_name = 1)

pd.read_excel("summer_raw.xls", sheet_name = "summer", skiprows= [0,1])

pd.read_excel("summer_raw.xls", sheet_name = "summer", skiprows= 2, usecols= "D:L")

summer = pd.read_excel("summer_raw.xls", sheet_name = "summer", skiprows= 2, usecols= "D:L")

summer.head()

summer.tail()

summer.info()

summer.to_csv("summer_imp.csv", index= False)

pd.read_csv("summer_imp.csv")

# summer.to_excel("summer_imp.xls") # old

summer.to_excel("summer_imp.xls", engine = "openpyxl", index= False) # new

pd.read_excel("summer_imp.xls")
