==> Udemy/PandasBootcamp/Course_Materials_Part2/Video_Lecture_NBs/NB_16_Seaborn_Basics.ipynb <==
# Part 2: Full Data Workflow A-Z

## Advanced / statistical plotting with Seaborn

### First Steps

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

titanic = pd.read_csv("titanic.csv")

titanic.head()

plt.figure(figsize=(12,8))
sns.set(font_scale=2, palette= "viridis")
sns.countplot(data = titanic, x = "sex", hue = "pclass")
plt.show()

## Categorical Plots

plt.figure(figsize=(12,8))
sns.set(font_scale=1.5)
sns.stripplot(data = titanic, x = "sex", y = "age", jitter = True, hue = "pclass", dodge = True)
plt.show()

plt.figure(figsize=(12,8))
sns.set(font_scale=1.5)
sns.swarmplot(data = titanic, x = "sex", y = "age", hue = "pclass", dodge = True)
plt.show()

plt.figure(figsize=(12,8))
sns.set(font_scale=1.5)
sns.violinplot(data = titanic, x = "sex", y = "age", hue = "pclass", dodge = True)
sns.swarmplot(data = titanic, x = "sex", y = "age", hue = "pclass", dodge = True, palette = 'dark:black') # new
plt.show()

plt.figure(figsize=(12,8))
sns.set(font_scale=1.5)
sns.violinplot(data = titanic, x = "pclass", y = "age", hue = "sex", dodge = True, split = True )
plt.show()

plt.figure(figsize=(12,8))
sns.set(font_scale=1.5)
sns.barplot(data = titanic, x = "pclass", y = "age", hue = "sex", dodge = True)
plt.show()

plt.figure(figsize=(12,8))
sns.set(font_scale=1.5)
sns.pointplot(data = titanic, x = "pclass", y = "age", hue = "sex", dodge = True)
plt.show()

## Jointplots / Regression

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

titanic = pd.read_csv("titanic.csv")

titanic.head()

sns.set(font_scale=1.5)
sns.jointplot(data = titanic, x = "age", y = "fare", height = 8, kind = "reg")
plt.show()

sns.set(font_scale=1.5)
sns.lmplot(data = titanic, x = "age", y = "fare", aspect= 1, height=8, col = "sex")
plt.show()

sns.set(font_scale=1.5)
sns.lmplot(data = titanic, x = "age", y = "survived", aspect= 1, height=8, col = "sex", logistic= True)
plt.show()

sns.set(font_scale=1.5)
sns.lmplot(data = titanic, x = "age", y = "survived", aspect= 1, height=8, col = "pclass", logistic= True)
plt.show()

## Matrixplots / Heatmaps

import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt

titanic = pd.read_csv("titanic.csv")

titanic.head()

pd.crosstab(titanic.sex, titanic.pclass)

plt.figure(figsize=(12,8))
sns.set(font_scale=1.4)
sns.heatmap(pd.crosstab(titanic.sex, titanic.pclass), annot= True, fmt = "d", cmap = "Reds", vmax = 150)
plt.show()

pd.crosstab(titanic.sex, titanic.pclass, values= titanic.survived, aggfunc= "mean")

plt.figure(figsize=(12,8))
sns.set(font_scale=1.4)
sns.heatmap(pd.crosstab(titanic.sex, titanic.pclass, values= titanic.survived, aggfunc= "mean"), annot= True, cmap = "Reds")
plt.show()

titanic.corr(numeric_only= True)

plt.figure(figsize=(12,8))
sns.set(font_scale=1.4)
sns.heatmap(titanic.corr(numeric_only=True), annot= True, cmap = "Reds")
plt.show()
